requirements.txt
boto3>=1.34,<2.0
Flask>=3.0,<4.0
python-dotenv>=1.0,<2.0

